+++ .vscode/settings.json
@@ -1,6 +1,5 @@
 {
 	"typescript.tsdk": "frontend\\node_modules\\typescript\\lib",
-	"eslint.workingDirectories": ["frontend", "backend"],
 	"cSpell.words": ["shadcn", "signin", "sonarjs", "Sonner", "stringifier"],
 	"terminal-keeper.killProcess": false,
 	"terminal-keeper.wslSupport": false,
