+++ backend/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "backend",
+	"name": "@z-commerce/backend",
 	"type": "module",
 	"version": "1.0.0",
 	"main": "src/server.ts",
@@ -15,21 +15,22 @@
 	},
 	"dependencies": {
 		"@colors/colors": "^1.6.0",
-		"@zayne-labs/toolkit": "^0.8.48",
-		"@zayne-labs/tsconfig": "^0.7.6",
+		"@z-commerce/shared": "workspace:*",
+		"@zayne-labs/toolkit": "^0.9.2",
+		"@zayne-labs/tsconfig": "^0.7.9",
 		"bcryptjs": "^3.0.2",
 		"consola": "^3.4.2",
 		"cookie-parser": "^1.4.7",
 		"cors": "^2.8.5",
 		"cross-env": "^7.0.3",
 		"date-fns": "^4.1.0",
-		"express": "5.0.1",
+		"express": "5.1.0",
 		"express-mongo-sanitize": "^2.2.0",
 		"express-rate-limit": "^7.5.0",
 		"helmet": "^8.1.0",
 		"hpp": "^0.2.3",
 		"jsonwebtoken": "^9.0.2",
-		"mongoose": "^8.12.1",
+		"mongoose": "^8.13.1",
 		"morgan": "^1.10.0",
 		"nodemailer": "^6.10.0",
 		"tsx": "^4.19.3",
@@ -43,9 +44,8 @@
 		"@types/hpp": "^0.2.6",
 		"@types/jsonwebtoken": "^9.0.9",
 		"@types/morgan": "^1.9.9",
-		"@types/node": "^22.13.11",
+		"@types/node": "^22.13.16",
 		"@types/nodemailer": "^6.4.17",
-		"@zayne-labs/prettier-config": "^0.7.6",
 		"typescript": "^5.8.2"
 	}
 }
