+++ eslint.config.js
@@ -4,14 +4,14 @@ export default zayne(
 	{
 		ignores: ["frontend/dist"],
 		react: true,
-		tailwindcss: {
-			overrides: {
-				"tailwindcss/no-unnecessary-arbitrary-value": "off", // Turned off cuz using a custom root font-size (10px)
-			},
-			settings: {
-				config: "frontend/tailwind.config.ts",
-			},
-		},
+		// tailwindcss: {
+		// 	overrides: {
+		// 		"tailwindcss/no-unnecessary-arbitrary-value": "off", // Turned off cuz using a custom root font-size (10px)
+		// 	},
+		// 	settings: {
+		// 		config: "frontend/tailwind.config.ts",
+		// 	},
+		// },
 		tanstack: {
 			query: true,
 		},
@@ -20,7 +20,7 @@ export default zayne(
 		},
 	},
 	{
-		files: ["backend/src/**/*"],
+		files: ["backend/src/**/*.ts"],
 		rules: {
 			"import/no-named-as-default-member": "off",
 		},
