+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "frontend",
+	"name": "@z-commerce/frontend",
 	"type": "module",
 	"version": "0.0.0",
 	"private": true,
@@ -10,21 +10,23 @@
 		"preview": "vite preview"
 	},
 	"dependencies": {
-		"@ark-ui/react": "^5.1.0",
+		"@ark-ui/react": "^5.4.0",
 		"@heroui/skeleton": "2.2.10",
 		"@heroui/theme": "2.4.12",
-		"@hookform/resolvers": "^4.1.0",
+		"@hookform/resolvers": "^5.0.0",
 		"@monicon/react": "^1.2.2",
 		"@monicon/vite": "^1.2.2",
-		"@tanstack/react-query": "^5.69.0",
-		"@zayne-labs/callapi": "1.6.14",
-		"@zayne-labs/toolkit": "^0.8.48",
+		"@tanstack/react-query": "^5.71.1",
+		"@z-commerce/shared": "workspace:*",
+		"@zayne-labs/callapi": "1.6.15",
+		"@zayne-labs/toolkit": "^0.9.2",
+		"@zayne-labs/ui-react": "^0.5.1",
 		"clsx": "^2.1.1",
-		"motion": "^12.4.7",
-		"react-hook-form": "^7.54.2",
+		"motion": "^12.6.2",
+		"react-hook-form": "^7.55.0",
 		"react-loading-skeleton": "^3.5.0",
-		"react-router": "^7.4.0",
-		"sonner": "^2.0.1",
+		"react-router": "^7.4.1",
+		"sonner": "^2.0.3",
 		"tailwind-merge": "2.6.0",
 		"tailwind-variants": "0.3.1",
 		"zod": "^3.24.2",
@@ -32,26 +34,24 @@
 	},
 	"devDependencies": {
 		"@hookform/devtools": "^4.3.3",
-		"@iconify/react": "^5.2.0",
-		"@tanstack/react-query-devtools": "^5.69.0",
+		"@iconify/react": "^5.2.1",
+		"@tanstack/react-query-devtools": "^5.71.1",
 		"@total-typescript/ts-reset": "^0.6.1",
-		"@types/node": "^22.13.11",
+		"@types/node": "^22.13.16",
 		"@types/react": "^19.0.12",
 		"@types/react-dom": "^19.0.4",
 		"@vitejs/plugin-react-swc": "^3.8.1",
-		"@zayne-labs/prettier-config": "^0.7.6",
 		"@zayne-labs/tsconfig": "^0.7.6",
 		"autoprefixer": "^10.4.20",
 		"postcss": "^8.5.3",
-		"prettier": "^3.5.1",
 		"prettier-plugin-classnames": "^0.7.6",
-		"prettier-plugin-merge": "^0.7.2",
+		"prettier-plugin-merge": "^0.7.3",
 		"prettier-plugin-tailwindcss": "^0.6.11",
-		"react": "19.0.0",
-		"react-dom": "19.0.0",
+		"react": "19.1.0",
+		"react-dom": "19.1.0",
 		"tailwindcss": "3.4.17",
 		"typescript": "^5.7.3",
-		"vite": "^6.1.1",
+		"vite": "^6.2.4",
 		"vite-tsconfig-paths": "^5.1.4"
 	}
 }
