--- frontend/src/components/primitives/For/index.ts
@@ -1,2 +1,0 @@
-export * as For from "./For";
-export { getElementList } from "./getElementList";
