+++ frontend/src/components/primitives/Slot.tsx
@@ -1,2 +1,2 @@
 /* eslint-disable react-refresh/only-export-components */
-export * from "@zayne-labs/ui-react/form";
+export * from "@zayne-labs/ui-react/common/slot";
