--- frontend/src/components/primitives/Slot/index.ts
@@ -1,1 +1,0 @@
-export { Slot, Slottable } from "./Slot";
