+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "z-commerce-root",
+	"name": "@z-commerce/root",
 	"type": "module",
 	"version": "1.0.0",
 	"packageManager": "pnpm@9.15.4",
@@ -13,13 +13,16 @@
 		"lint:format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@eslint-react/eslint-plugin": "^1.37.0",
+		"@eslint-react/eslint-plugin": "^1.40.0",
 		"@tanstack/eslint-plugin-query": "^5.66.1",
-		"@zayne-labs/eslint-config": "^0.7.6",
-		"eslint": "^9.22.0",
+		"@zayne-labs/eslint-config": "^0.7.9",
+		"@zayne-labs/prettier-config": "^0.7.9",
+		"@zayne-labs/tsconfig": "^0.7.9",
+		"eslint": "^9.23.0",
 		"eslint-plugin-react-hooks": "^5.1.0",
 		"eslint-plugin-react-refresh": "^0.4.19",
 		"eslint-plugin-tailwindcss": "^3.17.5",
+		"prettier": "^3.5.1",
 		"tailwindcss": "3.4.17"
 	}
 }
