+++ shared/package.json
@@ -1,0 +1,20 @@
+{
+	"name": "@z-commerce/shared",
+	"type": "module",
+	"version": "1.0.0",
+	"exports": {
+		".": null,
+		"./schemas": "./src/schemas/index.ts"
+	},
+	"scripts": {
+		"check-types": "tsc --noEmit",
+		"dev": "tsup --watch"
+	},
+	"dependencies": {
+		"zod": "^3.24.2"
+	},
+	"devDependencies": {
+		"@zayne-labs/tsconfig": "^0.7.9",
+		"typescript": "^5.8.2"
+	}
+}
